requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-classic
chromadb
pypdf
python-dotenv